Update to use gpg-based ssh authentication
Instead of relying on `ssh-agent`, rely on GnuPG's emulation of an ssh
agent, requiring that I only need a singular key for both signing and
ssh authentication.

Since this is implying that I will be signing each of my commits, and I
consistently forget how to organize my `.gitconfig`, write a markdown
document outlining how to make a basic `.gitconfig` suiting my needs.